feat: reestructure folder and files
