--- src/actionTypes.js
@@ -1,3 +1,0 @@
-export const ADD_TASK = "ADD_TASK"
-export const REMOVE_TASK = "REMOVE_TASK"
-export const COMPLETE_TASK = "COMPLETE_TASK"
