+++ src/store/configureStore.js
@@ -1,5 +1,5 @@
 import { legacy_createStore as createStore } from "redux";
-import reducer from "./reducer";
+import reducer from "./tasks";
 
 const store = createStore(reducer)
